small fixes

+++ main.py
@@ -158,7 +158,7 @@ class AudioUIApp(QtWidgets.QMainWindow, AudioUI.Ui_MainWindow):
         self.text_brows_info.append(f"Preparing file...")
         number_of_samples = round(len(data) * self.target_sample_rate / sample_rate)
         self.data_audio_file = sps.resample(data, number_of_samples, window="blackmanharris")
-        self.data_audio_file = self.data_audio_file.astype(np.int16)
+        # self.data_audio_file = self.data_audio_file.astype(np.int16)
         self.logger.debug(f"new sample rate: {self.target_sample_rate}")
         # else:
             # self.data_audio_file = data.astype(np.int16)
@@ -171,7 +171,8 @@ class AudioUIApp(QtWidgets.QMainWindow, AudioUI.Ui_MainWindow):
             self.data_audio_file = AudioUIApp.normalize(self.data_audio_file, max_val, target_max_val)
             # self.data_audio_file = AudioUIApp.butter_lowpass_filter(data=self.data_audio_file,
                                     # cutoff=6000, sample_rate=self.target_sample_rate, order=5)
-            self.data_audio_file = self.data_audio_file.astype(np.int16)
+                                    
+        self.data_audio_file = self.data_audio_file.astype(np.int16)
